Respect API/profile restrictions for extensions, if specified.
Affects mainly GLES and KHR extensions. For example, KHR_debug
extension names are without `KHR` suffix on desktop GL, but they do have
the suffix on ES.

For example, currently, if KHR_debug extension was requested either on
desktop or on ES, it resulted in both `glObjectLabel()` and
`glObjectLabelKHR()` being added to the generated files, which may be
confusing and error-prone to use and may cause the "get proc address"
code to complain that one of these functions is not found.

Another example is ARB_robustness on core desktop profile, which now
doesn't contain functions like glGetnHistogramARB() that were only part
of the compatibility profile. Sadly this is not true for
EXT_direct_state_access names, which are not properly marked in the
XML and thus still include ugly compatibility-profile-only names like
glMatrixPushEXT().

In short, this change should result in slightly shorter generated files
without irrelevant "noise".