Vulkan: dereference enum aliases only if the target isn't written yet.
For Vulkan 1.2, where the mirror clamp to edge extension was promoted,
the following change appeared:

     VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE = 4,
-    VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE_KHR = VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE
+    VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE_KHR = 4

Which doesn't look right as there are now two enums with the same value
but no apparent relation. Changing the code to dereference only if the
alias target isn't written yet makes it behave like before, showing the
relation that the _KHR-suffixed enum is an alias to the unsuffixed
(instead of the other way around or something).